0.10.6 (#26)
* 0.8.1 - move to OOP

* 0.8.1 - rewrite run_eda_tools to oop

* 0.8.2 - rewrite run_eda_tools to oop

* 0.8.3 - rewrite run_eda_tools to oop

* 0.8.4 - rewrite run_eda_tools to oop

* 0.8.5 - rewrite tcl_scr_gen to oop

* 0.8.6 - rewrite tcl_scr_gen to oop

* 0.8.7 - move vars from .tcl to vars_config.tcl

* 0.8.8 - change runtime view

* 0.8.9 - add mmmc preset to [MMMCGEN-2] message

* 0.8.10 - corr check_tables

* 0.8.11 - corr q3 answer in eun_eda_tools

* 0.8.12 - corr tcl_scr_gen.py

* 0.8.13 - corr tcl_scr_gen.py

* 0.8.14 - corr runtime form

* 0.8.15 - corr tcl_scr_gen.py

* 0.8.16 - corr run_eda_tools.py

* 0.8.17 - corr run_eda_tools.py

* 0.8.18 - add rm logs for -from_step

* 0.8.19 - add rm logs for -from_step

* 0.8.20 - add rm logs for -from_step

* 0.8.21 - add rm logs for -from_step

* 0.8.22 - fix run_eda_tools.py

* 0.8.23 - add -to_step option

* 0.8.24 - add -to_step option

* 0.8.25 - add -step option ; testing

* 0.8.26 - add -step option ; testing

* 0.9.1 - add -formal flow ; first try

* 0.9.2 - add -formal flow ; corr the_flow.sh

* 0.9.3 - add TF_TMPDIR to install script

* 0.9.4 - add [-config] option; first try

* 0.9.5 - add [-config] option; first try

* 0.9.6 - add [-config] option; corr norm exit

* 0.9.7 - add [-config] option; auto finding tf_var_<flow>.py files

* 0.9.8 - add [-config] option; auto finding tf_var_<flow>.py files

* 0.10.1.0 - change chmod to 755 (#16)

* 0.10.2 (#17)

* 0.10.2.0 - corr exit into q1

* 0.10.2.1 - add -f to exit func

* 0.10.3 check tf config (#18)

* 0.10.3.0 - add init_9 messages and check for tf_config

* 0.10.3.1 - corr init_9 message

* 0.10.4.0 - add check for /.* dirs during tf_config searching (#19)

* 0.10.5.0 - add perehod_na_novuyu_stroku kostil

* 0.10.5.1 (#24)

* Issue 23 (#25)

* 0.10.5.2.0

* 0.10.5.2.1

+++ core/the_flow_scripts.py
@@ -192,7 +192,8 @@ if __name__ == "__main__":
 
     if global_tf_vars.tf_update_run_dir_in_cfg:
         for i in os.scandir(global_tf_vars.tf_run_dir_in_cfg):
-            os.remove(i)
+            if i.name != 'vars_config.tcl':
+                os.remove(i)
         MmmcGen.run_mmmc_gen()
         PhyGen.run_phy_gen()
